+++ src/ionosphere/msis/io.jl
@@ -1,6 +1,6 @@
 using DelimitedFiles: readdlm, writedlm
 using Printf: @sprintf
-using Term: @bold, @underline
+using StyledStrings: @styled_str
 
 ############################################################################################
 # Looking for files
@@ -266,8 +266,7 @@ function save_msis_data(msis_data, parameters; verbose=true)
         write(f, "\n")
         writedlm(f, msis_data)
     end
-    verbose && println("File " * @bold("$filename") * " created under " *
-            @underline("$directory") * ".")
+    verbose && println(styled"File {bold:$filename} created under {underline:$directory}.")
 
     return fullpath
 end